Move most examples from UploadsController spec -> Upload spec
In order to keep the examples as similar as possible to the originals,
I've temporarily introduced a jsonify method in the Upload spec.
